Refactor Neo4j graph into its own class
- Remove LoadingContext
- Extract utility functions to object helper
